Debugger: Add stack trace window

+++ Source/Project64/N64System/Debugger/Debugger-StackTrace.cpp
@@ -12,9 +12,11 @@
 #include "stdafx.h"
 
 #include "DebuggerUI.h"
+#include "Symbols.h"
 
 CDebugStackTrace::CDebugStackTrace(CDebuggerUI* debugger) :
-CDebugDialog<CDebugStackTrace>(debugger)
+CDebugDialog<CDebugStackTrace>(debugger),
+m_EntriesIndex(0)
 {
 }
 
@@ -22,18 +24,43 @@ CDebugStackTrace::~CDebugStackTrace()
 {
 }
 
+void CDebugStackTrace::PushEntry(uint32_t routineAddress, uint32_t callingAddress)
+{
+	if (m_EntriesIndex < STACKTRACE_MAX_ENTRIES)
+	{
+		m_Entries[m_EntriesIndex] = { routineAddress, callingAddress };
+		m_EntriesIndex++;
+	}
+}
+
+void CDebugStackTrace::PopEntry()
+{
+	if (m_EntriesIndex > 0)
+	{
+		m_EntriesIndex--;
+	}
+}
+
+void CDebugStackTrace::ClearEntries()
+{
+	m_EntriesIndex = 0;
+}
+
 LRESULT CDebugStackTrace::OnInitDialog(UINT /*uMsg*/, WPARAM /*wParam*/, LPARAM /*lParam*/, BOOL& /*bHandled*/)
 {
 	DlgResize_Init();
-
-	m_List.SetExtendedListViewStyle(LVS_EX_FULLROWSELECT, 0);
-
+	
 	m_List.Attach(GetDlgItem(IDC_STACKTRACE_LIST));
-	m_List.AddColumn("Routine", 0);
-	m_List.AddColumn("Name", 1);
+	m_List.AddColumn("Caller", 0);
+	m_List.AddColumn("Routine", 1);
+	m_List.AddColumn("Name", 2);
+	
 
-	m_List.SetColumnWidth(0, 60);
-	m_List.SetColumnWidth(1, 100);
+	m_List.SetColumnWidth(0, 70);
+	m_List.SetColumnWidth(1, 70);
+	m_List.SetColumnWidth(2, 160);
+
+	m_List.SetExtendedListViewStyle(LVS_EX_FULLROWSELECT);
 
 	WindowCreated();
 	return TRUE;
@@ -41,7 +68,7 @@ LRESULT CDebugStackTrace::OnInitDialog(UINT /*uMsg*/, WPARAM /*wParam*/, LPARAM 
 
 LRESULT CDebugStackTrace::OnActivate(UINT uMsg, WPARAM wParam, LPARAM lParam, BOOL& bHandled)
 {
-	RefreshList();
+	Refresh();
 	return FALSE;
 }
 
@@ -73,32 +100,46 @@ LRESULT CDebugStackTrace::OnListDblClicked(NMHDR* pNMHDR)
 	return 0;
 }
 
-void CDebugStackTrace::RefreshList()
+void CDebugStackTrace::Refresh()
 {
-	vector<uint32_t>* stackTrace = m_Debugger->StackTrace();
+	if (!m_Debugger->Breakpoints()->isDebugging())
+	{
+		return;
+	}
+
+	SetWindowText(stdstr_f("Stack Trace (%d)", m_EntriesIndex).c_str());
 
 	m_List.SetRedraw(FALSE);
 	m_List.DeleteAllItems();
 	
-	int count = stackTrace->size();
+	CSymbols::EnterCriticalSection();
 
-	if (count > 4000)
+	for (int i = 0; i < m_EntriesIndex; i++)
 	{
-		count = 4000;
-	}
-
-	for (int i = 0; i < count; i++)
-	{
-		uint32_t address = stackTrace->at(i);
-
-		char szAddress[9];
-		sprintf(szAddress, "%08X", address);
+		uint32_t routineAddress = m_Entries[i].routineAddress;
+		uint32_t callingAddress = m_Entries[i].callingAddress;
 		
+		char szAddress[9];
+		sprintf(szAddress, "%08X", callingAddress);
 		m_List.AddItem(i, 0, szAddress);
-		m_List.AddItem(i, 1, "symbol");
 
-		m_List.SetItemData(i, address);
+		sprintf(szAddress, "%08X", routineAddress);
+		m_List.AddItem(i, 1, szAddress);
+
+		CSymbolEntry* symbol = CSymbols::GetEntryByAddress(routineAddress);
+		if(symbol != NULL)
+		{ 
+			m_List.AddItem(i, 2, symbol->m_Name);
+		}
+		else
+		{
+			m_List.AddItem(i, 2, "");
+		}
+		
+		m_List.SetItemData(i, routineAddress);
 	}
+
+	CSymbols::LeaveCriticalSection();
 
 	m_List.SetRedraw(TRUE);
 }
